services/video-generator/requirements.txt
# Video Generation Service Requirements
# Supports both CUDA and Metal (MPS) acceleration

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
safetensors>=0.4.0

# Video processing
opencv-python>=4.8.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# Image processing
Pillow>=10.0.0

# API server
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
numpy>=1.24.0
tqdm>=4.66.0
